Move evaluation into its own file.
